Extract structuredCloneTest
